Check isInteresting for methods and fields in ApiLint
The issue happens because we are running API lint when comparing
between new and old APIs surface which means that visitClass for
BigInteger does not get called (not a new class), but visitMethod
for byteValueExact does trigger. ApiLint skip check used to only
be called for ClassItem, which works for most cases, except for
new methods/fields in an old class.

Bug: 162035124
Test: New test added
Change-Id: Ia1be825fd9d51bdbb11b8d3f00edef44ddee8b59
Signed-off-by: Akash Srivastava <akashniki@gmail.com>